+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 **Changed:**
 - Compute Service - Updated default values for fault and update domains
+- Compute Service - Provided option to configure fault and update domains with reasonable default values
 
 ## 0.3.2